requirements.txt
autopep8==1.6.0
Babel==2.9.1
blinker==1.4
click==8.1.2
colorama==0.4.4
djlint==0.7.6
Flask==2.1.1
Flask-Babel==2.0.0
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-Sijax==0.4.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
future==0.18.2
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pathspec==0.9.0
Pillow==9.1.0
pycodestyle==2.8.0
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
regex==2022.3.15
Sijax==0.3.2
six==1.16.0
SQLAlchemy==1.4.35
toml==0.10.2
tomlkit==0.9.2
tqdm==4.64.0
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
